Add layouts folder and move headers and footer
