+++ client/src/components/layout/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import "../styles/footer.css"
+import "../../styles/footer.css"
 
 const Footer = () => {
