+++ client/src/pages/Login.js
@@ -1,6 +1,6 @@
 import React, {useContext, useState, useEffect, useRef} from 'react'
 import "../styles/edit.css"
-import PageHeader from '../components/PageHeader';
+import PageHeader from '../components/layout/PageHeader';
 import AuthContext from "../context/authContext"
 import Alerts from "../components/Alerts"
 import AlertContext from "../context/alertContext";
